Factor out token_source to test
